rename BaseTetreeKey -> TetreeKey.  TetreeKey -> ExtendedTetreeKey.  Reformat

+++ lucien/src/main/java/com/hellblazer/luciferase/lucien/NodeEstimator.java
@@ -68,7 +68,7 @@ public class NodeEstimator {
         long totalMemory = nodeCount * (baseMemoryPerNode + avgEntitiesPerNode * perEntityMemory);
 
         // Add overhead for sorted indices set
-        long sortedSetOverhead = nodeCount * 24; // TreeSet entry overhead
+        long sortedSetOverhead = nodeCount * 24L; // TreeSet entry overhead
 
         return totalMemory + sortedSetOverhead;
     }
@@ -107,7 +107,7 @@ public class NodeEstimator {
 
     private static float estimateTreeOverhead(int entityCount, int maxEntitiesPerNode, byte maxDepth) {
         // Estimate tree depth utilization
-        float avgEntitiesPerNode = (float) Math.min(entityCount / 10.0f, maxEntitiesPerNode * 0.7f);
+        float avgEntitiesPerNode = Math.min(entityCount / 10.0f, maxEntitiesPerNode * 0.7f);
         float depthUtilization = (float) (Math.log(entityCount / avgEntitiesPerNode) / Math.log(8));
         float normalizedDepth = Math.min(depthUtilization / maxDepth, 1.0f);
 
@@ -171,10 +171,10 @@ public class NodeEstimator {
                                                                                              0.1f);
         public static final SpatialDistribution SURFACE_THIN       = new SpatialDistribution(Type.SURFACE_ALIGNED, 0.3f,
                                                                                              1, 0.01f);
-        private final Type  type;
-        private final float clusteringFactor; // 0.0 = perfectly uniform, 1.0 = highly clustered
-        private final int   clusterCount;       // For CLUSTERED type
-        private final float surfaceThickness; // For SURFACE_ALIGNED type
+        private final       Type                type;
+        private final       float               clusteringFactor; // 0.0 = perfectly uniform, 1.0 = highly clustered
+        private final       int                 clusterCount;       // For CLUSTERED type
+        private final       float               surfaceThickness; // For SURFACE_ALIGNED type
 
         public SpatialDistribution(Type type) {
             this(type, 0.0f, 10, 0.1f);
@@ -202,6 +202,7 @@ public class NodeEstimator {
         public Type getType() {
             return type;
         }
+
         public enum Type {
             UNIFORM,        // Entities randomly distributed throughout space
             CLUSTERED,      // Entities grouped in clusters
